requirements.txt
numpy
torch>=1.7
opencv_python
loguru
scikit-image
tqdm
torchvision>=0.10.0
Pillow
thop
ninja
tabulate
tensorboard
motmetrics
filterpy
h5py
einops
pyyaml
argparse
easydict
numpy
tensorboardX
scipy
lap
cython
cython-bbox
fvcore
diffusers
transformers
accelerate
xformers
pandas